chore: rename data.ts to generate-data-by-month.ts
