+++ src/pages/ContentScripts/features/repo-fork-tooltip/view.tsx
@@ -6,7 +6,7 @@ import optionsStorage, {
   HypercrxOptions,
   defaults,
 } from '../../../../options-storage';
-import { generateDataByMonth } from '../../../../helpers/data';
+import generateDataByMonth from '../../../../helpers/generate-data-by-month';
 import ReactTooltip from 'react-tooltip';
 import ForkChart from './ForkChart';
